+++ config/environments/test.rb
@@ -1,5 +1,5 @@
 require "active_support/core_ext/integer/time"
-require_relative "../../test/lib/clients/oidc_test"
+require_relative "../../app/lib/utils/oidc_provider"
 
 # The test environment is used exclusively to run your application's
 # test suite. You never need to work with it otherwise. Remember that
@@ -67,10 +67,10 @@ Rails.application.configure do
   config.action_controller.raise_on_missing_callback_actions = true
 
   def configure_oidc
-    binding.irb
-    OIDCTest.configure_oidc_provider
+    provider = OidcProvider.new
+    provider.configure
     Clients::Vault.configure_oidc_client(config.astral.oidc_provider[:issuer],
-                                         OIDCTest.client_id,
-                                         OIDCTest.client_secret)
+                                         provider.client_id,
+                                         provider.client_secret)
   end
 end
